Refactor `SeedSettingsControl` and related controls
